reduce nested directories and name files according to component
