+++ src/cannon/collisions/COLLISION_GROUPS.js
@@ -1,0 +1,6 @@
+export const COLLISION_GROUPS = {
+  BALL: 1,
+  PLAYFIELD: 2,
+  FLIPPER: 4,
+  TRIGGER: 5
+}
