+++ src/cannon/collisions/bumperCollisionHandler.js
@@ -1,13 +1,5 @@
 import * as CANNON from 'cannon-es';
 
-export const COLLISION_GROUPS = {
-  BALL: 1,
-  PLAYFIELD: 2,
-  FLIPPER: 4,
-  TRIGGER: 5
-}
-
-
 export const bumperCollisionHandler = (e) => {
   const { 
     x:xBall,
