+++ src/cannon/materials/CONTACT_MATERIALS.js
@@ -1,0 +1,14 @@
+export const CONTACT_MATERIALS = {
+  ballAndPlayfield: {
+    friction: 0.1,
+    restitution: 0.6
+  },
+  ballAndBumper: {
+    friction: 0.0,
+    restitution: 0.5
+  },
+  ballAndFlipper: {
+    friction: 0.2,
+    restitution: 0.7
+  }
+}
