+++ src/cannon/materials/MATERIALS.js
@@ -1,0 +1,9 @@
+import * as CANNON from 'cannon-es';
+
+export const MATERIALS = {
+  ballMaterial: new CANNON.Material('ball'),
+  bumperMaterial: new CANNON.Material('bumper'),
+  flipperMaterial: new CANNON.Material('flipper'),
+  playFieldMaterial: new CANNON.Material('playfield')
+}
+
