--- src/cannon/triggers/index.js
@@ -1,1 +1,0 @@
-export { DrainTrigger } from './DrainTrigger';
