+++ src/debug/index.js
@@ -1,1 +1,3 @@
 export { DebugRenderer } from './DebugRenderer';
+export { InitKeyEvents } from './InitKeyEvents';
+export { DebugFloorMesh } from './DebugFloorMesh';
