+++ src/math/index.js
@@ -1,7 +1,1 @@
-import * as CANNON from 'cannon-es';
-
-export const getQuaternion = ({x, y, z}, radians) => {
-  const quaternion = new CANNON.Quaternion();
-  quaternion.setFromAxisAngle(new CANNON.Vec3(x, y, z), radians);
-  return quaternion;
-}
+export { getQuaternionFromAxisAngle } from './getQuaternionFromAxisAngle';
