+++ src/three/Camera.js
@@ -1,5 +1,5 @@
 import * as THREE from 'three';
-import { CAMERA } from '../../App.config';
+import { CAMERA } from '@src/App.config';
 
 export function Camera() {
   const camera = new THREE.PerspectiveCamera(75, window.innerWidth / window.innerHeight, 0.1, 1000);
