+++ src/three/Scene.js
@@ -14,7 +14,6 @@ export function Scene() {
     }
   }
 
-
   return {
     get self() { return scene },
     add
